+++ lib/src/intialize_connection.dart
@@ -1,6 +1,0 @@
-import 'dart:io';
-
-Future<HttpServer> initializeHttpConnection({required int port}) async {
-  final server = await HttpServer.bind('localhost', port, shared: true);
-  return server;
-}